pyproject.toml
[tool.poetry]
name = "agentsim"
version = "1.0.0"
description = "Modular agentic simulation framework for generating training traces"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{ include = "agentsim" }]

[tool.poetry.scripts]
agentsim = "agentsim.cli:main"

[tool.poetry.dependencies]
python = "^3.10"
pyyaml = "^6.0.1"
loguru = "^0.7.2"
python-dotenv = "^1.0.0"
httpx = "^0.27.0"
sentence-transformers = "^2.2.2"
rank-bm25 = "^0.2.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
black = "^23.11.0"
ruff = "^0.1.6"
matplotlib = "^3.10.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"

